Extract MockPrefsGlobalSession from MockGlobalSession.
